move custom types to central module, broaden scope for date_filter, move confidence_filter to viirs module, add OpenAPI tag groups
