+++ terraform/modules/storage/outputs.tf
@@ -5,3 +5,7 @@ output "tiles_bucket_domain_name" {
 output "tiles_bucket_website_endpoint" {
   value = aws_s3_bucket.tiles.website_endpoint
 }
+
+output "tiles_bucket_name" {
+  value = aws_s3_bucket.tiles.id
+}
